Benchmark: Enable single|multi node membership benchmark with a single script
Refactor stop/start scripts and unify usage
