+++ tests/start_server
@@ -14,38 +14,24 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-echo "export MULTINODE=N [2|4] env variable to execute the benchmark in replica mode. N number of folowers"
-echo "export PROFILING=true env variable to enable go profiling. Default profiling time is 60. Output: profile.svg"
+echo "export MULTINODE=true env variable to execute the benchmark in replica mode"
 
 echo Create id_ed25519 key
 ssh-keygen -t ed25519 -N '' -f /var/tmp/id_ed25519
 
-if [ ! -z "$PROFILING" ]
-then
-    echo PROFILING=enabled
-    PROFILING=--profiling
-    PROFILER_CMD="go tool pprof -svg -seconds 60 -output profile.svg http://localhost:6060"
-else
-    echo PROFILING=disabled
-    PROFILER_CMD="true"
-fi
-
 if [ -z "$MULTINODE" ]
 then
     echo Starting single server...
-    go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l info  --http-addr :8080 --raft-addr :9000 --mgmt-addr :8090 $PROFILING &
+    go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l info  --http-addr :8080 --raft-addr :9000 --mgmt-addr :8090 &
     sleep 10
-    $PROFILER_CMD &
     echo done.
 else
     echo Starting multinode...
-    go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l error --http-addr :8080 --raft-addr :9000 --mgmt-addr :8090 $PROFILING &
+    go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l error --http-addr :8080 --raft-addr :9000 --mgmt-addr :8090 &
     sleep 10
-    echo Starting followers
-    for i in $(seq 1 $MULTINODE); do
-        go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l error --http-addr :808$i --join-addr :8090 --raft-addr :900$i --mgmt-addr :809$i --node-id node$i &
-    done
+    echo Starting fallowers
+    go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l error --http-addr :8081 --join-addr :8090 --raft-addr :9001 --mgmt-addr :8091 --node-id node1 &
+    go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l error --http-addr :8082 --join-addr :8090 --raft-addr :9002 --mgmt-addr :8092 --node-id node2 &
     sleep 10
-    $PROFILER_CMD &
     echo done.
 fi
