+++ tests/stop_server
@@ -16,10 +16,11 @@
 
 echo Cleanup...
 if [ $(uname) == "Darwin" ]; then
-    pkill main
+    killall main || exit 0
 else
-    fuser -k -n tcp 8080
+    fuser -k -n tcp 8080 8081 8082 8083 8084 || true
 fi
+sleep 5
 rm -rf /var/tmp/demo.*
 rm -f /var/tmp/id_ed25519{,.pub}
 echo done.
