+++ tests/stress-membership
@@ -14,10 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+./start_server
 
-./start_server &
-sleep 5
-
-go run stress-membership.go
+go run riot.go
 
 ./stop_server
