--- tests/stress-multi-node-membership
@@ -1,41 +1,0 @@
-#!/usr/bin/env sh
-
-# Copyright 2018 Banco Bilbao Vizcaya Argentaria, S.A.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-echo Create id_ed25519 key
-ssh-keygen -t ed25519 -N '' -f /var/tmp/id_ed25519
-
-echo Starting multinode...
-go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l error --http-addr :8080 --raft-addr :9000 --mgmt-addr :8090 &
-sleep 5
-
-echo Starting fallowers
-go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l error --http-addr :8081 --join-addr :8090 --raft-addr :9001 --mgmt-addr :8091 --node-id node1 &
-go run ../main.go start -k pepe -p $(mktemp -d /var/tmp/demo.XXX) --raftpath $(mktemp -d /var/tmp/demo.XXX) -y /var/tmp/id_ed25519 -l error --http-addr :8082 --join-addr :8090 --raft-addr :9002 --mgmt-addr :8092 --node-id node2 &
-sleep 5
-echo done.
-
-go run stress-membership.go
-
-sleep 1
-echo Cleanup...
-if [ $(uname) == "Darwin" ]; then
-    killall main
-else
-    fuser -k -n tcp 8080 8081 8082
-fi
-rm -rf /var/tmp/demo.*
-rm -f /var/tmp/id_ed25519*
-echo done.
